backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
pydantic-settings==2.6.1
sqlalchemy==2.0.34
sqlmodel==0.0.21
alembic==1.13.3
psycopg[binary]==3.2.13
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
